# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestoría Canarias — Boletines Oficiales
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * BOE, BOC, BOP_LPA, BOP_TFE, or BORME
 */
export type EntrySource = typeof EntrySource[keyof typeof EntrySource];


export const EntrySource = {
  BOE: 'BOE',
  BOC: 'BOC',
  BOP_LPA: 'BOP_LPA',
  BOP_TFE: 'BOP_TFE',
  BORME: 'BORME',
} as const;

export interface Entry {
  id: number;
  /** BOE, BOC, BOP_LPA, BOP_TFE, or BORME */
  source: EntrySource;
  title: string;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  category?: string | null;
  publishedAt: string;
  url: string;
  isRead: boolean;
  isBookmarked: boolean;
  createdAt: string;
}

export interface EntryList {
  entries: Entry[];
  total: number;
  page: number;
  limit: number;
}

export type SyncResultSourcesItem = {
  source?: string;
  count?: number;
  success?: boolean;
};

export interface SyncResult {
  success: boolean;
  message: string;
  newEntries: number;
  sources?: SyncResultSourcesItem[];
}

export interface SyncStatus {
  /** @nullable */
  lastSyncAt: string | null;
  isRunning: boolean;
  /** @nullable */
  lastResult?: string | null;
  /** Cron schedule expression */
  scheduledTime?: string;
}

export type StatsSummaryBySourceItem = {
  source?: string;
  count?: number;
};

export interface StatsSummary {
  totalEntries: number;
  unreadCount: number;
  todayCount: number;
  bookmarkedCount: number;
  bySource: StatsSummaryBySourceItem[];
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface Alert {
  id: number;
  category: string;
  /** @nullable */
  source?: string | null;
  createdAt: string;
}

export interface CreateAlertBody {
  category: string;
  source?: string;
}

export interface AlertMatch {
  alert: Alert;
  entries: Entry[];
}

export interface Note {
  id: number;
  entryId: number;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateNoteBody {
  content: string;
}

export type ClientNifType = typeof ClientNifType[keyof typeof ClientNifType] | null;


export const ClientNifType = {
  nif: 'nif',
  cif: 'cif',
} as const;

export type ClientType = typeof ClientType[keyof typeof ClientType] | null;


export const ClientType = {
  autonomo: 'autonomo',
  sl: 'sl',
  sa: 'sa',
  asociacion: 'asociacion',
  comunidad_propietarios: 'comunidad_propietarios',
  otro: 'otro',
} as const;

export type ClientTaxRegime = typeof ClientTaxRegime[keyof typeof ClientTaxRegime] | null;


export const ClientTaxRegime = {
  modulos: 'modulos',
  estimacion_directa: 'estimacion_directa',
  estimacion_directa_simplificada: 'estimacion_directa_simplificada',
  otro: 'otro',
} as const;

export interface Client {
  id: number;
  name: string;
  /**
     * NIF (personas físicas) o CIF (personas jurídicas)
     * @nullable
     */
  nif?: string | null;
  nifType?: ClientNifType;
  bormeMonitored: boolean;
  type?: ClientType;
  /** @nullable */
  cnae?: string | null;
  /** @nullable */
  municipality?: string | null;
  taxRegime?: ClientTaxRegime;
  keywords?: string[] | null;
  active: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateClientBodyType = typeof CreateClientBodyType[keyof typeof CreateClientBodyType];


export const CreateClientBodyType = {
  autonomo: 'autonomo',
  sl: 'sl',
  sa: 'sa',
  asociacion: 'asociacion',
  comunidad_propietarios: 'comunidad_propietarios',
  otro: 'otro',
} as const;

export type CreateClientBodyTaxRegime = typeof CreateClientBodyTaxRegime[keyof typeof CreateClientBodyTaxRegime];


export const CreateClientBodyTaxRegime = {
  modulos: 'modulos',
  estimacion_directa: 'estimacion_directa',
  estimacion_directa_simplificada: 'estimacion_directa_simplificada',
  otro: 'otro',
} as const;

export interface CreateClientBody {
  name: string;
  /** NIF/NIE/CIF español válido — se infiere nif_type automáticamente */
  nif?: string;
  type?: CreateClientBodyType;
  cnae?: string;
  municipality?: string;
  taxRegime?: CreateClientBodyTaxRegime;
  keywords?: string[];
  active?: boolean;
  bormeMonitored?: boolean;
}

export type UpdateClientBodyType = typeof UpdateClientBodyType[keyof typeof UpdateClientBodyType];


export const UpdateClientBodyType = {
  autonomo: 'autonomo',
  sl: 'sl',
  sa: 'sa',
  asociacion: 'asociacion',
  comunidad_propietarios: 'comunidad_propietarios',
  otro: 'otro',
} as const;

export type UpdateClientBodyTaxRegime = typeof UpdateClientBodyTaxRegime[keyof typeof UpdateClientBodyTaxRegime];


export const UpdateClientBodyTaxRegime = {
  modulos: 'modulos',
  estimacion_directa: 'estimacion_directa',
  estimacion_directa_simplificada: 'estimacion_directa_simplificada',
  otro: 'otro',
} as const;

export interface UpdateClientBody {
  name?: string;
  /** NIF/NIE/CIF español válido — se infiere nif_type automáticamente */
  nif?: string;
  type?: UpdateClientBodyType;
  cnae?: string;
  municipality?: string;
  taxRegime?: UpdateClientBodyTaxRegime;
  keywords?: string[];
  active?: boolean;
  bormeMonitored?: boolean;
}

export type EntryClientMatchMatchedBy = typeof EntryClientMatchMatchedBy[keyof typeof EntryClientMatchMatchedBy];


export const EntryClientMatchMatchedBy = {
  rules: 'rules',
  ai: 'ai',
} as const;

export interface EntryClientMatch {
  id: number;
  entryId: number;
  clientId: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  relevanceScore: number;
  /** @nullable */
  reason?: string | null;
  matchedBy: EntryClientMatchMatchedBy;
  reviewed: boolean;
  createdAt: string;
}

export type EntryClientMatchWithClientMatchedBy = typeof EntryClientMatchWithClientMatchedBy[keyof typeof EntryClientMatchWithClientMatchedBy];


export const EntryClientMatchWithClientMatchedBy = {
  rules: 'rules',
  ai: 'ai',
} as const;

export interface EntryClientMatchWithClient {
  id: number;
  entryId: number;
  clientId: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  relevanceScore: number;
  /** @nullable */
  reason?: string | null;
  matchedBy: EntryClientMatchWithClientMatchedBy;
  reviewed: boolean;
  createdAt: string;
  client: Client;
}

export type CreateMatchBodyMatchedBy = typeof CreateMatchBodyMatchedBy[keyof typeof CreateMatchBodyMatchedBy];


export const CreateMatchBodyMatchedBy = {
  rules: 'rules',
  ai: 'ai',
} as const;

export interface CreateMatchBody {
  clientId: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  relevanceScore: number;
  reason?: string;
  matchedBy: CreateMatchBodyMatchedBy;
}

export type ListEntriesParams = {
/**
 * Filter by bulletin source (BOE, BOC, BOP_LPA, BOP_TFE)
 */
source?: string;
/**
 * Filter by category
 */
category?: string;
/**
 * Filter by publication date (YYYY-MM-DD)
 */
date?: string;
/**
 * Full-text search in title and summary
 */
search?: string;
/**
 * Page number (1-indexed)
 */
page?: number;
/**
 * Items per page
 */
limit?: number;
};

export type GetRecentEntriesParams = {
limit?: number;
};

export type ListClientsParams = {
/**
 * Filter by active status
 */
active?: boolean;
};

export type UpdateEntryMatchBody = {
  reviewed?: boolean;
  /**
     * @minimum 1
     * @maximum 3
     */
  relevanceScore?: number;
  reason?: string;
};