# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Gestoría Canarias — Boletines Oficiales
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Bad request
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Client:
      type: object
      required:
        - id
        - name
        - active
        - bormeMonitored
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        nif:
          type: ["string", "null"]
          description: "NIF (personas físicas) o CIF (personas jurídicas)"
        nifType:
          type: string
          enum: [nif, cif]
          nullable: true
        bormeMonitored:
          type: boolean
        type:
          type: string
          enum: [autonomo, sl, sa, asociacion, comunidad_propietarios, otro]
          nullable: true
        cnae:
          type: ["string", "null"]
        municipality:
          type: ["string", "null"]
        taxRegime:
          type: string
          enum: [modulos, estimacion_directa, estimacion_directa_simplificada, otro]
          nullable: true
        keywords:
          type: array
          items:
            type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateClientBody:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        nif:
          type: string
          description: "NIF/NIE/CIF español válido — se infiere nif_type automáticamente"
        type:
          type: string
          enum: [autonomo, sl, sa, asociacion, comunidad_propietarios, otro]
        cnae:
          type: string
        municipality:
          type: string
        taxRegime:
          type: string
          enum: [modulos, estimacion_directa, estimacion_directa_simplificada, otro]
        keywords:
          type: array
          items:
            type: string
        active:
          type: boolean
        bormeMonitored:
          type: boolean
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        nif:
          type: string
          description: "NIF/NIE/CIF español válido — se infiere nif_type automáticamente"
        type:
          type: string
          enum: [autonomo, sl, sa, asociacion, comunidad_propietarios, otro]
        cnae:
          type: string
        municipality:
          type: string
        taxRegime:
          type: string
          enum: [modulos, estimacion_directa, estimacion_directa_simplificada, otro]
        keywords:
          type: array
          items:
            type: string
        active:
          type: boolean
        bormeMonitored:
          type: boolean
...